R_1, photodiode updated, Test_pin added for opt.
The diode was flipped the wrong way (doh)

R_1 was changed to 200 kOhm as this gave finer results on lower voltages
(LCDs), still not decided where this should be or if potentiometer
should be added.

Test_pin added in order to detect computing times for certain tasks.
Last update before removing UART.